Address PR comments / checks.
